Respond to float to double changes.

+++ apps/wire-cell-2dtoy.cxx
@@ -25,7 +25,7 @@ int main(int argc, char* argv[])
   WireCellSst::GeomDataSource gds(argv[1]);
   WireCell2dToy::FrameDataSource fds(10, gds);
 
-  std::vector<float> ex = gds.extent();
+  std::vector<double> ex = gds.extent();
   cerr << "Extent: "
        << " x:" << ex[0]/units::mm << " mm"
        << " y:" << ex[1]/units::m << " m"
